[i2c] Monitor stretching and lost arbitration for control symbols
The control symbols were overlooked for clock stretching and
arbitration. SCL can be stretched on (repeated) Start and Stop symbols
like any other, and in the multi-controller context, SCL can be pulled
low early as well. If SCL is pulled low before the controller intended,
treat it as clock synchronization if SDA had already reached its desired
level first, and treat it as lost arbitration if SDA did not.

Categorize any SCL pulse that is shorter than the round-trip time as a
runt pulse. Add SCL interference events for runt SCL pulses detected by
the bus monitor, and treat these as arbitration loss, since the SCL
movement is too fast to satisfy any of the timing requirements.

Signed-off-by: Alexander Williams <awill@opentitan.org>