[hw,mbx,rtl] Move SOC_INTR_ADDR/DATA registers to the end
A mailbox con either be a PCIe mailbox or a FW mailbox. In case of the
PCIe setting, the SOC reagisters 0x0 and 0x4 implement the capability
header functionality, but this is done outside of the mbx RTL. PCIe
mailboxes do not use the SOC_INTR_DATA/ADDR registers. FW-based
mailboxes on the other hand only use the SOC_INTR_DATA/ADDR registers
but no capability headers. Previously, it was decided to alias both
registers as they are used orthogonally. However, this deceission was
revised.

This PR moves the SOC_INTR_DATA/ADDR registers to the end of the SOC
register interface to not alias with any capability header registers
for PCIe mailboxes

Signed-off-by: Robert Schilling <rschilling@rivosinc.com>