[sival,sram_ctrl] Fix sram_ctrl_sleep_ret_contents test
This test needed a workaround for a bug in the ROM/ROM_EXT in which the
ROM_EXT expects a valid `version` to be present in the retention SRAM in
order to boot.

This version is lost when scrambling.

To work around it, we have had to swap around the order of the normal
(due to watchdog) reset and the deep sleep reset. The watchdog reset
reverts back to the original key when it next boots, while the deep
sleep reset preserves the new scrambling key.

We can only correctly store / restore the `version` field with this
ordering, not the original.

This is fixed in the production version of the ROM.

Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>