[csrng/rtl] Move cmd checks to the cmd stage
Having the command checks in the main SM doesn't make a lot of sense.
This causes a lot of issues including a hang condition inside the cmd stage
in case of generate commands with cmd len greater than 1. For this reason
it is best to not even start processing commands that violate our checks.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>