[manuf test] fix fake CA CSR and add a script for generating the CA cert
The set of fake_ca* files allows the user to generate a self signed CA
certificate which can be used for validating of test generated TPM
certificates. This patch modifies the configuration to ensure that the
contents of the Issuer field of the CA certificate matches the TPM and
UDS certificates' expectations.

A convenience script is being added to automate generation of the fake
CA ceritificate in DER and PEM forms.

This has been tested on top of the entire patchset, 'openssl verify'
command confirms validity of TPM certificates.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
(cherry picked from commit fb4d5c62b404c7eab966d24c0821475b42e6b1fe)