[silicon_creator,flash_ctrl] Prepare for saving TPM certificates
All three TPM certificates will be saved in the same freshly allocated
for this purpose flash INFO page.

This page needs to be included in the list of the pages prepared for
personalization.

Tested when the on top of the entire patch stack, TPM certificates can
be saved and rertrieved from the assigned flash INFO page.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>