[personalization] add generation of TPM certificates
TPM certificate templates are added to the set, the certificates are
added to the serialization structures as necessary.

TPM objects received over serdes have some truncating bytes, which are
now being discarded, but when calculating the number of flash words
necessary to save the certificate one extra word is added if the
certificate size is not evenly divisible by 4 (the flash word size).

All three TPM certificates are saved in the same flash INFO page. Note
that the base offsets of the certificates must be 8 bytes aligned due
to flash controller limitations.

Tested on the FPGA setup by successfully running
 bazel run //sw/host/opentitantool -- --interface=hyper310 fpga \
   clear-bitstream && \
 bazel test --test_output=streamed \
 //sw/device/silicon_creator/manuf/skus/earlgrey_a0/sival_bringup:ft_provision_fpga_cw310_sival

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>