[rv_dm,dv] Tidy up halt_resume_whereto vseq
This sequence is currently failing, after some changes I made to the
DMI driver/monitor a few weeks ago. It turns out that the problem is
caused by the fact that the vseq was doing a DMI operation by directly
driving the JTAG signals (instead of using the DMI driver).

Fixing this was reasonably easy but I realised that the sequence
itself didn't give much confidence about what it was doing. This
commit replaces magic hex constants and adds lots of comments to
explain what it is actually doing. The behaviour should be mostly
unchanged.

The hex constants were a little tricky to interpret! Indeed, figuring
out the exact abstract command that was running in this sequence was a
bit hard: we were just writing the hex constant 'h5C00021102 to the
"dmi" register.

I decoded this with the RISC-V debug specification. Firstly (section
6.1.5), the constant splits into fields as

    address = 'h17
    data    = 'h00008440
    op      = 2'b10 = 2

op = 2 means the operation is "write data to address". address = 'h17
is "command" (the abstract command register)

The abstract command register is documented in section 3.12.7 of the
same specification and has fields cmdtype, control. The 'h8440
constant splits up as:

    cmdtype = 8'h00
    control = 24'h008440

A cmdtype of zero means "Access register command". Then constant in
control is decoded as

    aarsize          = 2
    aarpostincrement = 0
    postexec         = 1
    transfer         = 0
    write            = 0
    regno            = 16'h40 (CSR 64)

After this commit, the code should do the same thing but it will
hopefully be a bit easier to understand.

The only extra change that I've made is to add an extra busy check to
the end of the vseq, which should make this all a bit more
symmetrical. I've added the check to the testpoint to keep things in
sync.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>