[keymgr] Don't block PRNG reseeding operations while processing commands
A code comment mentioned that this was due to DV limitations but there
are no signs that this limitation still exists. Removing this
qualification of the seed enable signal is more efficient from an
entropy consumption viewpoint as the entropy is anyway requested but
just dropped in case of commands being processed currently.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>