[usb,host] Add a USB harness in rust
This harness, partially copied from the sival branch, is responsible
for bootstrapping, waiting until the OT USB device appears and then
can launch an arbitrary executable to interact with the USB device.
The harness monitors the UART output of the device and stops the test
whenever the device reports a passing status. The harness indicates
to the executable which device to use using a command line switch
`--device <bus>:<addr>`. It is possible to pass extra arguments by
using `--exec-arg=...`

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>