[usb,host] Improve USB harness
This commit gives the harness the capability to run a binary after the
OT USB device is detected.
The harness monitors the UART output of the device and stops the test
whenever the device reports a passing status. The harness indicates
to the executable which device to use using a command line switch
`--device <bus>:<addr>`. It is possible to pass extra arguments by
using `--exec-arg=...`

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>
(cherry picked from commit 057859f2617fb1c7cd8b631e01faa56a470bdb60)