[sw/tests] Add missing initialization and reset to `entropy_src_edn_reqs_test`
Before this commit, this test initialized keymgr with an invalid Creator
Root Key from OTP, because the relevant OTP partition did not get
locked.  This did not cause problems because keymgr currently does not
raise an error and go into the Invalid state.  This behavior of keymgr
will be changed/fixed in a subsequent commit, so this commit fixes the
`entropy_src_edn_reqs_test` so that it doesn't fail when keymgr gets
fixed.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>