[jtag,dv] Do not collect JTAG items with zero DR and IR length
The driver can sometimes decide to send special "dummy" DR or IR
transactions. This is reasonable, but the end result is that we fall
out of the Update-DR or Update-IR state in the FSM without any actual
data.

This is hard to represent in our data model because jtag_item
distinguishes between the two cases (when something has come from the
monitor) by expecting exactly one of ir_len and dr_len to be positive.

To avoid a dummy packet where both lengths are zero (which isn't even
recognisable as IR or DR), tweak the monitor so that it no longer
creates the transaction item.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>