[dv] Add more prints to bit bash sequence
I was just debugging why a particular invocation was dying (for
rv_dm) with a timeout. It's a bit tricky to see whether something has
got stuck or whether everything is just a bit slow.

To make things a bit easier to debug, I've added a print (at
UVM_MEDIUM) that gives a bit more information about how far we've got.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>