[rv_dm,dv] Double timeout for rv_dm_jtag_dmi_csr_bit_bash
This test is sometimes timing out. The reason is that I tweaked the
DMI driver to deal more gracefully with JTAG frequencies much higher
than the clock frequency. The (usually overly careful) code we ended
up with is a bit slower and we take more than 50ns to run through the
whole list with some seeds.

Double the time available. The simulation that was timing out still
runs in only 3m20s on my laptop so I'm not really concerned about
total time.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>