[csrng/dv] Add fork to command transaction collection
This commit adds a fork to the command transaction collection.
This could come in handy when commands are aborted by CSRNG due to
a command error status response. Otherwise the collection might
wait for the genbits to be received even if no genbits will
ever arrive for the command at hand.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>