[jtag,dv] Remove exit_to_rti_ir machinery
This was used to control whether the IR path ran to
completion (getting back to Run-Test/Idle). The code sort of worked,
but it wasn't actually fully wired up: if we had a req that left us in
UpdateIr (instead of RTI), the following req wouldn't actually behave
accordingly, because we weren't reading exit_to_rti_ir_past (which I
deleted in the previous commit).

Getting this exactly right will be a bit fiddly, and the code that's
here is not really going to help. Drop it and do it properly when we
need to.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>