[edn/dv] This commit fixes the deadlock in edn_disable_auto_req_mode
Currently there is a deadlock in the edn_disable_auto_req_mode test.
This deadlock occurs whenever the EDN is disabled during the initialization
of auto mode. The config variable cfg.abort_sw_cmd is used to signal to
the base vseq that the EDN will be disabled. However, this variable is not
taken into consideration in all of the apropriate places. Furthermore,
the variable is not being looked at during spinwaits. For this reason
reset_asserted() needs to be called to tell the csr_utils_pkg that there
is an incoming re-enablement (even though we are not resetting and this
still works and I prefer this option over adding an additional function to
csr_utils_pkg for re-enablements).
This commit makes sure that for edn_disable_auto_req_mode, the EDN is only
disabled, when no csr acceses are outstanding (which if IIUC is the intention
behind cfg.abort_sw_cmd).

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>