[opentitantool] Advanced bitbanging
This CL enhances the existing FTDI-inspired bitbanging supported by
HyperDebug.  In particular, it adds means of specifying that generation
of the waveform should be put on "pause" at a certain point, until a
certain logic level is seen on particular pins.

This feature allows generating I2C waveforms, while allowing the I2C
device to stretch the clock.  Or generating an edge on a GPIO pin at a
very precise time after OT has toggled some other pin, in order for a
test case to attempt to trigger a race condition.

Change-Id: Ie42a35b87d961047a2ed4c10bc7d6f5ec5a89b92
Signed-off-by: Jes B. Klinke <jbk@chromium.org>