[manuf] cleanup `ft_personalize`
This cleans up the `ft_personalize` binary to consolidate operations
into sub-functions, and increase readability. Additionally this fixed an
initialization bug in the size of the TPM TBS certificate buffers.
Lastly, this updates the certificate hash comparison on the host side to
print both hashes if the comparison fails (so that it is easier to see
how the host interpretted the bytes, including in the order of said
bytes).

Signed-off-by: Tim Trippel <ttrippel@google.com>