[jtag,dv] Handle trst_n at start of time in jtag_driver.sv
The existing code worked as long as the get_and_drive() task started
after we had driven trst_n to zero AND we didn't ever want to drive it
to zero afterwards.

It turns out that this wasn't the situation I was seeing in my
simulations and the result was that the driver thought the JTAG FSM
was in the RTI state, but it was actually in Test-Logic-Reset and
everything got rather confused.

With this change, we will always notice if we see a trst_n reset, and
the driver and bus shouldn't get out of sync in the same way.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>