[dv,clkmgr] Generate clkmgr binds
Generate the sva and coverage binds so they can support multiple tops.
The coverage and sva binds are specific to top_earlgrey.
Remove unused ports from clkmgr_aon_cg_en_sva_if.

The pwrmgr bind also needs to be changed to deal with this, but is
not generated (yet).

Fixes #19448

Signed-off-by: Guillermo Maturana <maturana@opentitan.org>