[manuf]: add code to create Cloud KMS signed CA certificate
The modification makes use of the openssl ability to interact with
Google Cloud KMS when the appropriate configuration set.

When the appropriate environment variables are set to point to a local
instance of Google KMS integration library
(https://github.com/GoogleCloudPlatform/kms-integrations) and Cloud
KMS configuration, the second certificate, signed by the configured
external crypto engine is also generated.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>