[rv_dm,dv] Slightly weaken check in rv_dm_halt_resume_whereto_vseq
This check works by reading from the WHERETO magic address. We expect
to get a jump instruction (jal), but it turns out that there are two
possible destinations: the start of the program buffer and the start
of an abstract command.

Rather than modelling exactly which destination to expect, let's
just weaken the check to ensure we get a jump to one of them.

This fixes a failure mode for the stress sequence where we were
failing because the rv_dm_cmderr_busy_vseq had written something to
the command register before rv_dm_halt_resume_whereto_vseq ran, which
meant we ended up pointing at the abstract command instead of the
program buffer.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>