[keymgr] Don't update and reseed PRNG in Disabled/Invalid state forever
Previously, keymgr would keep updating and reseeding the PRNG forever
once entering the StCtrlDisabled or StCtrlInvalid state. This is not
ideal from an entropy and power consumption viewpoint.

This commit changes the design to - once one of the two states is
entered - to keep updating the PRNG (which also triggers the reseed
operation) until two more PRNG reseed operation have happened.

This resolves lowRISC/OpenTitan#22997.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>