[jtag,dv] Sync up with TCK before calling drive_jtag_req
This fixes the occasional seed for rv_dm_stress_all. The seq_item_port
that get_and_drive samples from isn't always full. It normally gets a
new item on a clock edge, but the clock edge is probably for the main
clock in the testbench: not the jtag TCK.

The result is that we enter drive_jtag_req at some undetermined point
in the TCK cycle, and that clock might happen to be running, so the
code that was at the start of tms_tdi_step() doesn't necessarily
trigger to get us back in sync.

This change will sort that out, jumping to the next HOST_CB event at
the start of get_and_drive().

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>