[runtime,log] Make logging usable in DIFs
Presently, it is impossible to use the logging macros in DIFs
for somewhat obscure reasons. This is because the DIFs are compiled
on host for opentitanlib bindgen, and it fails with a really difficult
to debug error about signing select having wrong values. On the other
hand, being able to print from the DIFs is really useful when debugging.

This commit change the log library so that it can compile on host, hence
enabling usage in DIFs. It requires a bit of ugly conditional compilation
because logging in DV relies on magic symbols.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>