[sw] Fix SAFETY comments in `status.rs`
This commit moves the safety comments above what they're annotating.

It also adds a safety comment about the `transmute` from bytes loaded
from an ELF into an extern C struct.

Had to extract `mod_id_ptr` to change the `is_err_status` line's
formatting to fit on one line. This is to work around a clippy bug
which is fixed in
<https://github.com/rust-lang/rust-clippy/pull/11170>
but our Rust toolchain is too old.

Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>