[edn/dv] Add transition to reject entropy state to alert test
This commit adds transitions from each state to the reject entropy state
to the alert vseq.
Furthermore, there is now added functionality for sending responses to the
EDN without the EDN making any request. This way the coverage will still
be recorded which wouldn't be the case if we just force signals.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>