[opentitantool] Non-blocking bitbanging trait
In order to make best use of the new advanced bit-banging support, one
should be able to perform other operations with opentitantool and
HyperDebug, while a bit-banging waveform is waiting for a trigger
condition.

This CL makes it such that while the `opentitantool gpio bitbang ...`
command still returns only when the waveform is complete, the underlying
opentitanlib methods are changed such that they must be repeatedly
polled.  This works well with `opentitansession`, in that the session
daemon process being single-threaded will be able to respond to the
repeated queries, while also serving requests from other invocations of
`opentitantool` in between.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>

Change-Id: Ic51166db1bae30eed31f9adda162966c9cd01fff