[i2c,dv] Rework the target_runtime_base_vseq for upcoming tests
This vseq defines a test framework that was used for some of the
fifo reset tests. This framework was useful when scoreboard modelling
was insufficient to correctly predict these sorts of async behaviours.

This change cleans and refactors the test framework into something
much more principled, clean and flexible for implementing upcoming
directed test sequences. The test sequences that previously derived from
this class have been adapted to continue operating as before.

This commit also cleans up an adjacent interrupt handler routine in the same file.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>