[flash_ctrl] Data XOR address infection
This commit adds the data XOR address infection mechanism. The goal of
this countermeasure is to detect faults on read address targeting the flash
controller.

On a flash read, the address used for accessing the data is XORed inside the
`flash_phy_rd` module to the fetched data.

The address then is removed using an XOR at these points before an integrity check
is conducted:
- Inside `u_tl_adapter_eflash` where the core has a TL-UL interface to the flash memory.
  Here, the buffered address from the incoming TL-UL read request is taken.
- Before the data goes into the `u_to_rd_fifo` where reads from SW are returned over TL-UL.
  Here, the buffered address from the incoming TL-UL read request is taken.
- Inside `u_flash_hw_if` where hardware components (life cycle, key manager, and OTP)
  have access to the flash.

When a FI attack manipulates an address, data would be fetched from the wrong address in
the flash memory. But inside the `flash_phy_rd` module, this manipulated address is XORed
to the data. Now, when removing the address by using the original address, and there is a
difference from a FI, the data would be faulty and the subsequent integrity check detects
the attack.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>