[i2c,dv] Change target interrupt handler to check pins sequentially
The previous priority-decoder method of handling active wired
interrupts does not jive with status-type interrupts, unless we
wish to use the testbench to mask interrupts via the enable csr.

Instead, give each interrupt's handler routine a chance to run, though
in a sequential order if multiple are pending.

Add declarations for yet-unimplemented handlers.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>