[rv_dm,dv] Connect mon_jtag_if to internal signals
This won't have much effect on the enable/disable events, but
it *does* change the timing because the decision about whether to
connect some of the signals goes through a prim_lc_sync or a
prim_clock_mux2, which injects a few extra cycles of delay.

Rather than trying to guess the exact timing, let's cheat and tell the
jtag monitor to use the version that's actually making it through to
the internals of the block.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>