[hmac/rtl] Wait for digest of complete block when stopping
Prior to this commit, when HMAC was told to stop, it would wait for the
completion of a complete block based on its message length counter.
However, the relevant message length counter for this is inside
`prim_sha2` and depends on the digest size, leading to problems such as
issue #23163.

This commit fixes these problems by letting `prim_sha2` signal when it
is computing a digest for a complete block and by taking that signal
into account in `hmac`'s `done_state` FSM.

This resolves #23163.

Co-authored-by: Andreas Kurth <adk@lowrisc.org>
Signed-off-by: Martin Velay <mvelay@lowrisc.org>