[cryptolib] Also poll CSRNG command ready bit for additional data
Previously, the command register ready bit was only polled for the
additional data when writing commands to the SW register of EDN.

When writing commands to CSRNG instead, not polling the command ready
bit could cause CSRNG to drop additional data provided by software while
handling a hardware command. Once the hardware command was handled, the
software instance would win arbitration and CSRNG would wait for
software to provide the dropped additional data, thereby stalling the
entire entropy complex. This would happen for example if a HW instance
request would collide with a software request. Since the software
command FIFO is only 2 entries deep, it can only be guaranteed to absorb
2 words if the software interface doesn't win arbitration immediately.

This is related to lowRISC/OpenTitan#22209 and lowRISC/OpenTitan#22210.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>